Used Maven Lifecycle tests, ашÑfixed whitespaces.
