+++ pom.xml
@@ -22,7 +22,7 @@
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
-            <version>4.12</version>
+            <version>4.13.2</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
@@ -64,5 +64,8 @@
             </plugins>
         </pluginManagement>
     </build>
+
+
+
 </project>
